+++ src/components/cart/index.js
@@ -1,0 +1,3 @@
+export { default as CartProductCard } from 'components/cart/CartProductCard/CartProductCard';
+export { default as CartProductList } from 'components/cart/CartProductList/CartProductList';
+export { default as OrderContainer } from 'components/cart/OrderContainer/OrderContainer';
