+++ src/components/common/Header/Header.jsx
@@ -4,7 +4,7 @@ import useWindowsSize from 'hooks/useWindowSize';
 
 import { ROUTE } from 'route';
 
-import Icon from 'components/common/Icon/Icon';
+import { Icon } from 'components/common';
 import * as Styled from 'components/common/Header/Header.style';
 
 import { deviceSizeStandard } from 'styles/Theme';
