+++ src/components/product/index.js
@@ -1,0 +1,3 @@
+export { default as CartAdd } from 'components/product/CartAdd/CartAdd';
+export { default as ProductCard } from 'components/product/ProductCard/ProductCard';
+export { default as ProductList } from 'components/product/ProductList/ProductList';
