+++ src/pages/Home/Home.jsx
@@ -1,8 +1,7 @@
 import React from 'react';
 
-import PageTemplate from 'components/common/PageTemplate/PageTemplate';
-import ProductList from 'components/product/ProductList/ProductList';
-import Pagination from 'components/common/Pagination/Pagination';
+import { PageTemplate, Pagination } from 'components/common';
+import { ProductList } from 'components/product';
 
 import * as Styled from 'pages/Home/Home.style';
