+++ src/pages/ProductDetail/ProductDetail.jsx
@@ -1,6 +1,5 @@
 import React from 'react';
-import ErrorContainer from 'components/common/ErrorContainer/ErrorContainer';
-import PageTemplate from 'components/common/PageTemplate/PageTemplate';
+import { ErrorContainer, PageTemplate } from 'components/common';
 
 function ProductDetail() {
   return (
